catch errors for log_action

+++ salor-retail/app/models/salor_base.rb
@@ -126,8 +126,13 @@ module SalorBase
     SalorBase.log_action(self.class.to_s,txt)
   end
   def self.log_action(from="unk",txt)
-    File.open("#{::Rails.root.to_s}/log/#{Rails.env}-history.log","a") do |f|
-      f.write "[#{Time.now}] [FROM: #{from}] " + txt + "\n"
+    return if not File.exists? "#{::Rails.root.to_s}/log/#{Rails.env}-history.log"
+    begin
+      File.open("#{::Rails.root.to_s}/log/#{Rails.env}-history.log","a") do |f|
+        f.write "[#{Time.now}] [FROM: #{from}] " + txt + "\n"
+      end
+    rescue
+      return
     end
   end
   def self.string_to_float(str)
